+++ app/src/main/java/com/example/mitrais/onestopclick/view/edit_movie/EditMovieViewModel.java
@@ -58,7 +58,7 @@ public class EditMovieViewModel extends AndroidViewModel {
         return productRepo.saveMovieUri(id, uri);
     }
 
-    Task<Void> saveRating(String id, HashMap<String, Integer> rating) {
+    Task<Void> saveRating(String id, HashMap<String, Float> rating) {
         return productRepo.saveRating(id, rating);
     }
